requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp==20211102.0
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cors==0.7.0
aiosignal==1.2.0
ansicon==1.89.0
arrow-cpp==14.0.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
aws-c-auth==0.6.19
aws-c-cal==0.5.20
aws-c-common==0.8.5
aws-c-compression==0.2.16
aws-c-event-stream==0.2.15
aws-c-http==0.6.25
aws-c-io==0.13.10
aws-c-mqtt==0.7.13
aws-c-s3==0.1.51
aws-c-sdkutils==0.1.6
aws-checksums==0.1.13
aws-crt-cpp==0.18.16
aws-sdk-cpp==1.10.55
blas==1.0
blessed==1.20.0
boost-cpp==1.82.0
bottleneck==1.3.7
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.8.30
cachetools==5.3.3
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorful==0.5.4
colorlog==5.0.1
contourpy==1.3.0
cryptography==43.0.0
cudatoolkit==11.6.0
cycler==0.12.1
dgl==2.2.1
distlib==0.3.8
et-xmlfile==1.1.0
faiss-cpu==1.8.0.post1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.0
fsspec==2024.6.1
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gflags==2.2.2
glog==0.5.0
gmpy2==2.1.2
google-api-core==2.19.1
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gpustat==1.1.1
grpc-cpp==1.48.2
grpcio==1.48.2
h5py==3.12.1
huggingface-hub==0.25.0
icc_rt==2022.1.0
idna==3.7
intel-openmp==2023.1.0
interchange==2021.0.4
jieba==0.42.1
jinja2==3.1.4
jinxed==1.2.0
joblib==1.4.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keras==3.6.0
kiwisolver==1.4.7
libansicon==1.89
libblas==3.9.0
libboost==1.82.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libclang==18.1.1
libcurl==8.9.1
libevent==2.1.12
libffi==3.4.4
liblapack==3.9.0
libprotobuf==3.20.3
libssh2==1.11.0
libthrift==0.15.0
libuv==1.48.0
lightgbm==4.5.0
llvmlite==0.43.0
lz4-c==1.9.4
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markdown==3.4.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mkl-service==2.4.0
ml-dtypes==0.4.1
monotonic==1.6
mpc==1.1.0
mpfr==4.0.2
mpir==3.0.0
mpmath==1.3.0
msgpack-python==1.0.3
msys2-conda-epoch==20160418
multidict==6.0.4
namex==0.0.8
networkx==3.3
ninja==1.11.1
numba==0.60.0
numexpr==2.8.4
numpy==1.24.3
nvidia-ml-py==12.535.133
opencensus==0.11.3
opencensus-context==0.1.3
openpyxl==3.1.5
openssl==3.1.0
opt-einsum==3.4.0
optree==0.13.0
orc==1.7.4
packaging==24.1
pandas==2.2.2
pansi==2020.7.3
pillow==10.4.0
pip==24.2
platformdirs==2.5.2
plotly==5.22.0
pooch==1.8.2
prometheus_client==0.13.1
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.0
py-spy==0.3.14
py2neo==2021.2.4
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11-abi==5
pycparser==2.21
pydantic==1.10.12
pygments==2.18.0
pyopenssl==24.2.1
pyparsing==3.1.4
pysocks==1.7.1
python==3.10.14
python-dateutil==2.9.0post0
python-tzdata==2023.3
python_abi==3.10
pytorch==1.12.1
pytz==2024.1
pyyaml==6.0.1
ray-core==2.6.3
ray-default==2.6.3
ray-tune==2.6.3
re2==2022.04.01
recbole==1.2.0
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.10.6
rsa==4.7.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.1.1
setproctitle==1.2.2
setuptools==72.1.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
smart_open==5.2.1
snappy==1.2.1
sqlite==3.45.3
sympy==1.13.2
tabulate==0.9.0
tbb==2021.8.0
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.0
tensorboardx==2.6.2.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
texttable==1.6.4
thop==0.1.1-2209072238
threadpoolctl==3.5.0
tk==8.6.14
tokenizers==0.19.1
torch-geometric==2.6.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torchdata==0.8.0
tqdm==4.66.5
transformers==4.44.2
typing-extensions==4.11.0
typing_extensions==4.11.0
tzdata==2024a
ucrt==10.0.22621.0
urllib3==2.2.2
utf8proc==2.6.1
vc==14.40
virtualenv==20.17.1
vs2015_runtime==14.40.33807
wcwidth==0.2.5
werkzeug==3.0.3
wheel==0.44.0
win_inet_pton==1.1.0
wrapt==1.16.0
xgboost==2.1.1
xz==5.4.6
yaml==0.2.5
yarl==1.11.0
zlib==1.2.13
zstd==1.5.5
